Fix index.html and finish two recipe pages
The Sour Cherry Pie recipe wasn't linked properly on the
index page, which was then fixed. The layout of the no-bean_chili
page was copied to the other two recipe pages. Theses pages
were then completed with information fitting the respective recipes.
A minor change consisted of the universal naming of the heading of
each page, which now look like "Dish X Recipe".